refactor($step3): change package name from step3 to lotto
