+++ src/main/java/lotto/domain/LottoTicketGenerator.java
@@ -1,4 +1,4 @@
-package step3.domain;
+package lotto.domain;
 
 import java.util.ArrayList;
 import java.util.Collections;
