+++ src/test/java/lotto/domain/ProfitCalculatorTest.java
@@ -1,8 +1,8 @@
-package step3.domain;
+package lotto.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import step3.common.WinningType;
+import lotto.common.WinningType;
 
 import static org.assertj.core.api.Assertions.*;
