+++ src/test/java/lotto/domain/WinningNumbersTest.java
@@ -1,9 +1,9 @@
-package step3.domain;
+package lotto.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import step3.common.ErrorCode;
-import step3.common.WinningType;
+import lotto.common.ErrorCode;
+import lotto.common.WinningType;
 
 import java.util.Arrays;
 import java.util.List;
